Improve release process for the Windows detached sig process
This makes the process more consistent with the OS X process. There are
other changes that should be made with the process, but these are just the
factual changes with the process as it stands right now, as well as the
change that can be made purely with the instructions.

[skip ci]