net: use dummy addresses for dns seed source groups
Rather than running a second dns lookup to get an ip that may change, set aside
a dummy ip range to represent these sources. This is needed for async dns
lookups, otherwise ugly queuing or polling is needed while awaiting the second
result.

A dummy range is used rather than passing in an explicit group hash because the
source address is serialized to disk, and a change would break back-compat.