net: remove one-connect-at-a-time hack
With the old synchronous connections, the connected groups were enough to make
sure that we weren't connecting multiple times to the same node. Now that we're
async, it's possible for addrman to hand us the same address twice before it's
marked as tried.

To fix, add connection attempts as well as current connections to the group,
and evict on connection failure or disconnect.