net: don't send any messages before handshake or after requested disconnect
In the event that the version handshake did not complete successfully (bad
version, unexpected services, etc), the peer's version is known. So rather than
using pnode->nVersion to decide if we should process/send their messages, use
pnode->fSuccessfullyConnected instead.

Also, send reject messages earlier in SendMessages(), so that disconnections are
processed earlier.

These changes combined should ensure that no message is ever sent after
fDisconnect is set (with the exception of feelers, which will be addressed as a
follow-up).