time: Use std::chrono for time rather than boost
Unfortunately, there's still no standard way of printing the current time in a
threadsafe way. Digging down into boost's approach, they simply use gmtime_r
when possible, as guessed by availability macros.

We now use the same approach, but use autotools to detect whether gmtime_r or
gmtime_s can be used, or as a fallback, the racy gmtime.

Note that MilliSleep was not replaced because it is an interruption point. That
can be done once boost threads are all gone.