net: add lock for closing a node's socket
Rather than trying to lock around the socket itself, add a new bool and lock
that's used when disconnecting the socket. This represents a barrier around the
disconnection itself.

This means that the socket can become const, since we can check to see if the
socket is already closed, rather than setting the socket value itself to
invalid.