temp hack: add network fuzzer
1. un-comment the changes in net.cpp
2. Run on testnet/mainnet for a while as usual. Net messages will be dumped to disk.
3. re-comment net.cpp
4. Run bitcoin_netfuzz

bitcoin_netfuzz will repeatedly connect to a local node, maintaining 8
connections. It will randomly replay messages from the supplied file. Now and
then, it will truncate the messages and/or insert the wrong checksum.

The only attempt made at satisfying the p2p protocol is that it will usually
(but not always) send a version message first.