test: Default SIGCHLD in DataDirectory test
Revert SIGCHLD to default during the test, otherwise boost.test will
catch and fail on it, when the testing subprocess ends: there is
BOOST_TEST_IGNORE_SIGCHLD but that only works when defined at build-time
of the boost library.

This is necessary to make the test pass on OpenBSD, but not Linux,
because `sig_info->si_code` is set differently.