squashme: Needs AMD benchmark. don't use intel-specific optimization
The algorithm used to calculate Sigma0/Sigma1 here is the same as the one used
in #13400, which resulted in a ~5% speedup on Intel CPUs, but close to 25%
penalty on AMD.

So unfortunately, this should result in a ~5% slowdown on Intel.

However, #13400 was operating on 128bit registers and these are 32bit, it's
possible that there's no penalty on AMD, and this change is not needed.