span: remove default copy ctor and assignment oper
These have the effect of disabling the default move constructor and move
operator. This is generally meaningless for spans due to their simplicity,
but consider the (simplified here) internal implementation of .first():

Span<C> first(std::size_t count) const noexcept {
    return Span<C>(m_data, count);
}

Technically this creates a new Span object and moves/copies it when it returns,
but RVO takes care of optimizing that out.

Because trivial moves have been implicitly deleted, it falls into the more
complicated conversion rvalue constructor to handle the return.

The next commit will add an attribute to the conversion constructor to help
detect dangling stack references, so we want to make sure that we're not
sending trivial copies through it.