Integrate univalue into our buildsystem
Requires https://github.com/bitcoin-core/univalue/pull/19 .

This addresses issues like the one in #12467, where some of our compiler flags
end up being dropped during the subconfigure of Univalue. Specifically, we're
still using the compiler-default c++ version rather than forcing c++11.

We can drop the need subconfigure completely in favor of a tighter build
integration, where the sources are listed separately from the build recipes,
so that they may be included directly by upstream projects. This is
similar to the way leveldb build integration works in Core.

Upstream benefits of this approach include:

    Better caching (for ex. ccache and autoconf)
    No need for a slow subconfigure
    No more missing compile flags
    Compile only the objects needed

There are no benefits to Univalue itself that I can think of. These changes
should be a no-op there, and to downstreams as well until they take advantage
of the new sources.mk. Once merged, This Core branch is ready-ish for PR, and
takes advantage of the features added here.

This also removes the option to use an external univalue to avoid similar ABI
issues with mystery binaries.