Bubble up fatal errors
Rather than calling abort in surprising places, pass up fatal errors and let
the application logic deal with them.

maybe_fatal_t can return an error message or a return value, but not both. This
is a deliberate design that ensures that a caller can never receive a real
return value if a fatal error was encountered.