build: Create .la library for crc32c
Libtool will yell at you if you try to link a shared library against
static ones.

This change creates a libtool archive library for crc32c and allows a
shared library to be linked against it portably.

Also append LIBTOOL_APP_LDFLAGS (basically -all-static) since we want
everything internally to be compiled the same. If we don't specify this,
then libtool will prefer the non-static PIC version of the object, which
is built with -DDLL_EXPORT -DPIC for mingw-w64 targets. This can cause
symbol resolution problems when we link this library against an
executable that does specify -all-static, since that will be built
without the -DDLL_EXPORT flag.

[META] This change is done in preparation for a future commit where we
       link the libbitcoinkernel library against this one.