Dont require bitcoin-config.h for serialize.h
Add yet another level of indirection to endian conversion

Clients can't use our endian definitions/macros because they rely on our
bitcoin-config.h (which is generated as part of our internal build process).

When building with BUILD_BITCOIN_INTERNAL and bitcoin-config.h present, the
_internal functions behave exactly as before.

When building without those defines present, the user gets a compiled-in
version of the same code.

This way the client is not required to generate a bitcoin-config.h or define
any variables for serialization to just work.

Note that all core objects should be built with BUILD_BITCOIN_INTERNAL, which
is not currently true.