mempool: POC: replace boost multi_index with custom structures
This is a hack and slash first attempt to verify that these structures are
suitable. They seem to be.

The multi_index is replaced with:
- An unordered_map of txids to CTxMemPoolEntry iterators
- An unordered_map of wtxids to CTxMemPoolEntry iterators
- An ordered (by fee rate) set of iterators to CTxMemPoolEntry
- An ordered (by entry time) set of iterators to CTxMemPoolEntry
- An ordered (by ancestor fee rate) set of iterators to CTxMemPoolEntry
- An unordered_map of CTxMemPoolEntry -> to a list of iterators to set items

Any time the main unordered_map is updated, the circular list of iterators is
refreshed.

Enough is implemented here for libbitcoinkernel to compile and link without
boost. I haven't attempted to run tests, it is DEFINITELY broken in lots and
lots of ways.

I made no attempt at clean code, and several other approaches were tried before
finding one that worked so there are probably remnants of that.