test_bitcoin_consensus POC
Add a new binary which exists purely to run tests on the external interface of
libbitcoin_consensus without linking concerns.

Note that this does not eliminate the need for testing the shared libs
themselves. For ex, proper exports still need to be verified.

This binary is not very interesting, but a similar one for libbitcoin_kernel
likely would be.

Issues/TODO:

- A dummy test fixture is used rather than dragging in our test util lib
- Redundant tests run on external/internal apis were removed. Maybe not
  worth worrying about?
- Duplicated BuildCreditingTransaction/BuildSpendingTransaction. Find them a
  better home?