POC: cache mpgen output and reuse it when possible
This is quick hack to demonstrate the concept. It is _very_ quick and dirty and
probably buggy.

This adds pre-generated source files from mpgen to the repo and uses them when
possible.

How it works:
.hash files are added for each generate file. These files contain the hash of
the source file which create them.

For each .capnp file, CMake checks to see if:
- every necessary generated file exists
- A hash file exists for each generated file
- The hash matches the actual hash of the source .capnp file

If so, it copies the committed source files to the build dir and uses them. If
any of those checks fail, mpgen is invoked to regenerate the files into the
build dir.

Any time the source files are touched, the above process re-runs so that the
typical "edit + make" process works as usual.

Currently the process to add the generated files and corresponding hash files
is manual. A convenience target could be added to automate it.

A big improvement would be for mpgen to embed the source hash in all generated
files instead, which would negate the  need for the extra .hash files.