net: don't hold m_nodes_mutex for long-lived callbacks
The shared_ptrs now ensure that the CNodes haven't been deleted for the
duration of the callbacks.

Disconnection/Finalization may happen while the shared_ptr is being held,
though it won't be deleted for the duration of the callback.

The majority of uses are in ProcessMessages and SendMessages, during which
disconnection may occur but FinalizeNode will not be called.