net: optimization: allow nodes to be finalized before they're disconnected
There is no reason to force a destruction dependency order between net and
net_processing.

Now that FinalizeNode() is performed on the message handler thread, it can be
used to clean up nodes immediately after a pass through the loop.

This means that nodes may be finalized before they're disconnected, but that's
no problem.